Updated Post Function
This update allows the post details to be stored in a JSON file to be later uploaded to a feed displaying these posts. I also made the code to save a draft of the post as a comment, as I am unsure if we will still be using this feature.